bump to v38.0.0

+++ tests/Driver/Datagram/PulseWidthEncoderTest.cs
@@ -9,7 +9,7 @@ public class PulseWidthEncoderTest
         var autd = CreateController();
 
         var rnd = new Random();
-        var buf = Enumerable.Range(0, 256).Select(_ => new PulseWidth((uint)rnd.Next(0, 512))).ToArray();
+        var buf = Enumerable.Range(0, 256).Select(_ => new PulseWidth((ushort)rnd.Next(0, 512))).ToArray();
 
         autd.Send(new AUTD3Sharp.PulseWidthEncoder(dev => i => buf[i.Item1]));
         foreach (var dev in autd)
@@ -25,7 +25,7 @@ public class PulseWidthEncoderTest
         var autd = CreateController();
 
         var buf = Enumerable.Range(0, 256).Select(i =>
-            new PulseWidth((uint)Math.Round(Math.Asin((float)i / 255) / MathF.PI * 512))
+            new PulseWidth((ushort)Math.Round(Math.Asin((float)i / 255) / MathF.PI * 512))
         ).ToArray();
 
         autd.Send(new AUTD3Sharp.PulseWidthEncoder());
